+++ src/Factory/Factory.php
@@ -1,0 +1,10 @@
+<?php
+
+
+namespace PlugHttp\Factory;
+
+
+interface Factory
+{
+    public static function create();
+}