+++ src/Factory/JsonFactory.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace PlugHttp\Factory;
+
+use PlugHttp\Body\Json;
+
+class JsonFactory implements Factory
+{
+    public static function create()
+    {
+        return new Json();
+    }
+}